fix: deepcopy filliing nil data with zeroes (#310)
Corrects the bug introduced by
https://github.com/celestiaorg/rsmt2d/pull/100. Currently, if
`deepcopy()` encounters nil inner slice values, it creates a slice of 0
length, which is incorrect. This PR addresses the issue. The problem
becomes evident when eds is partially computed, resulting in Row() or
Col() methods returning invalid copies.

On the side note, it is technically breaking public API change. But the
nature of the bug suggests noone relied on existing misbehaviour before.